fix(start-core): length-check the long description, not the short one twice
`Description::validate` checks `self.short` against 160, then checks
`self.short` again against 5000 under an error that says "Long description must
be 5000 characters or less". `self.long` is never looked at. Both fields are
`LocaleString`, so it compiles either way.

Note this changes nothing on its own: `Description::validate` has no callers.
`S9pk::validate_and_filter` -> `Manifest::validate_for` validates the archive's
contents, not the manifest's field values, and nothing else calls it — so
neither the 160 nor the 5000 limit is enforced anywhere today. Wiring it in is
a separate decision with an install-time blast radius, so it is deliberately
not done here.

For whoever takes that decision: nothing in either Start9 registry would fail
it. Across every package in start9-registry and community-registry the longest
short description in any locale is 124 characters and the longest long
description is 1713 (bitcoin-knots 29.x).

Not compiled — the change is one field access of identical type, and building
start-core from cold is a heavy build.

+++ shared-libs/crates/start-core/src/s9pk/v2/manifest.rs
@@ -432,7 +432,7 @@ impl Description {
                 crate::ErrorKind::ValidateS9pk,
             ));
         }
-        if match &self.short {
+        if match &self.long {
             LocaleString::Translated(s) => s.len() > 5000,
             LocaleString::LanguageMap(map) => map.values().any(|s| s.len() > 5000),
         } {
